+++ crates/dslab-dag/Cargo.toml
@@ -16,3 +16,5 @@ serde_yaml = "0.8"
 enum-iterator = "0.7.0"
 rand = "0.8.4"
 rand_pcg = "0.3.1"
+itertools = "0.10.5"
+threadpool = "1.8.1"
