+++ crates/dslab-dag/src/lib.rs
@@ -3,6 +3,7 @@
 pub mod dag;
 pub mod dag_simulation;
 pub mod data_item;
+pub mod experiment;
 pub mod network;
 pub mod parsers;
 pub mod resource;
